Revert "msm_fb: display: fix compiler warning for uint64_t format"
This reverts commit 415bb26fce8608f03685c3de0fa43f56a7760bf8.

Revert "msm: msm_bus: Add support for 64-bit bandwidth requests"

This reverts commit 13aba771e3672dd8909a84395f19ae719ebfc468.

+++ arch/arm/mach-msm/include/mach/msm_bus.h
@@ -43,8 +43,8 @@
 struct msm_bus_vectors {
 	int src; /* Master */
 	int dst; /* Slave */
-	uint64_t ab; /* Arbitrated bandwidth */
-	uint64_t ib; /* Instantaneous bandwidth */
+	unsigned int ab; /* Arbitrated bandwidth */
+	unsigned int ib; /* Instantaneous bandwidth */
 };
 
 struct msm_bus_paths {
